@@ -374,5 +374,5 @@ func (t *regionTree) TotalWriteRate() (bytesRate, keysRate float64) {
 }
 
 func init() {
-	rand.Seed(time.Now().UnixNano())
+	rand.New(rand.NewSource(time.Now().UnixNano()))
 }
