@@ -120,5 +120,5 @@ func NewWaitingOperatorStatus() *WaitingOperatorStatus {
 }
 
 func init() {
-	rand.Seed(time.Now().UnixNano())
+	rand.New(rand.NewSource(time.Now().UnixNano()))
 }
