@@ -15,7 +15,7 @@ func TestRegionPath(t *testing.T) {
 	f := func(id uint64) string {
 		return path.Join(regionPathPrefix, fmt.Sprintf("%020d", id))
 	}
-	rand.Seed(time.Now().Unix())
+	rand.New(rand.NewSource(time.Now().Unix()))
 	for i := 0; i < 1000; i++ {
 		id := rand.Uint64()
 		re.Equal(f(id), RegionPath(id))
